+++ tradingbot/order_type.py
@@ -1,0 +1,13 @@
+"""Trading operation class."""
+from enum import Enum
+
+
+class OrderType(str, Enum):
+  """Define the different operations in trading."""
+
+  BUY = 'buy'
+  SELL = 'sell'
+  BUYLIMIT = 'buylimit'
+  SELLLIMIT = 'selllimit'
+  BUYSTOP = 'buystop'
+  SELLSTOP = 'sellstop'