+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## ** V0.4.0  (11/02/2024) **
+- feat:
+  - strategy abstract class
+  - common useful methods in trading
+  - new methods in forex_client: get_bid_ask, place_break_even and get_pip
+- test:
+  - strategy abstract class
+  - common useful methods in trading
+
 ## ** V0.3.0  (25/01/2024) **
 - feat:
   - forex main to handle incoming historic data